+++ src/gocept/__init__.py
@@ -1,3 +1,1 @@
-# Copyright (c) 2010 gocept gmbh & co. kg
-# See also LICENSE.txt
-__import__('pkg_resources').declare_namespace(__name__)
+__import__('pkg_resources').declare_namespace(__name__)  # pragma: no cover